worked on user section
