+++ .env
@@ -17,5 +17,4 @@ NEXT_PUBLIC_AMEX_INVOICE_CANCEL_URL = https://school.mangoitsol.com/admin/invoic
 NEXT_PUBLIC_AMEX_SALES_ORDER_REDIRECT_URL = https://school.mangoitsol.com/admin/sales_order/list/sale
 NEXT_PUBLIC_AMEX_SALES_ORDER_CANCEL_URL = https://school.mangoitsol.com/admin/sales_order/list/sale
 
-
 NEXT_PUBLIC_WEBHOOKS_NOTIFICATIONS_SECRET=*INSERT_YOUR_WEBHOOK_NOTIFICATION SECRET_HERE*
