+++ pages/_app.tsx
@@ -10,6 +10,8 @@ import "../styles/addinvoice.css";
 import "../styles/globals.css";
 import React, { useState } from "react";
 import Footer from "./commoncmp/footer";
+import { useRouter } from "next/router";
 export default function App({ Component, pageProps }: AppProps) {
+  const route = useRouter()
   return <Component {...pageProps} />;
 }
