+++ pages/admin/customer/addNewParent.tsx
@@ -23,18 +23,13 @@ import {
   Stack,
 } from "@mui/material";
 import { useForm, SubmitHandler } from "react-hook-form";
-import { api_url, auth_token } from "../api/hello";
-import AddCustomerCmp from "../commoncmp/addCustomerCmp";
+import { api_url, auth_token } from "../../api/hello";
 import axios from "axios";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import CloseIcon from "@mui/icons-material/Close";
 import styled from "@emotion/styled";
-const style = {
-  color: "red",
-  fontSize: "12px",
-  fontWeight: "bold",
-};
+import { GridCloseIcon } from "@mui/x-data-grid";
+import commmonfunctions from "../../../commonFunctions/commmonfunctions"
 
 //dialog box
 const BootstrapDialog = styled(Dialog)(({ theme }) => ({}));
@@ -59,12 +54,18 @@ function BootstrapDialogTitle(props: DialogTitleProps) {
             color: (theme) => theme.palette.grey[500],
           }}
         >
-          <CloseIcon />
+          <GridCloseIcon />
         </IconButton>
       ) : null}
     </DialogTitle>
   );
 }
+
+const style = {
+  color: "red",
+  fontSize: "12px",
+  fontWeight: "bold",
+};
 
 interface TabPanelProps {
   children?: React.ReactNode;
@@ -90,7 +91,6 @@ function TabPanel(props: TabPanelProps) {
     </div>
   );
 }
-
 function a11yProps(index: number) {
   return {
     id: `simple-tab-${index}`,
@@ -100,74 +100,90 @@ function a11yProps(index: number) {
 
 type FormValues = {
   name: string;
-  status: number;
-  customertype: number;
   email1: string;
   email2: string;
-  number: number;
+  phone1: number;
+  phone2: number;
   contactName: string;
   printUs: string;
+  status: number;
+  type: number;
   parentId: number;
-  agegroup: number;
+  userRole: String;
   pregeneratedid: string;
-  address1: string;
-  address2: string;
-  city: string;
-  state: string;
-  postalcode: number;
-  attentionto: number;
-  alternatenumber: number;
 };
 
-export default function EditCustomer({
-  id,
+export default function AddNewParent({
   open,
-  closeDialogedit,
+  closeDialog,
+  name,
 }: {
-  id: any;
   open: any;
-  closeDialogedit: any;
+  closeDialog: any;
+  name: any;
 }) {
-  const [selvalue, setselValue] = React.useState(0);
+  const [value, setValue] = React.useState(0);
   const [spinner, setshowspinner] = React.useState(false);
-  const [opens, setOpen] = React.useState(open);
-  const [parentid, setparentid] = React.useState(0);
-  const [parentname, setparentname] = React.useState<any>("");
-  const [custtype, setcusttype] = React.useState<any>([]);
   const [btnDisabled, setBtnDisabled] = React.useState(false);
-  const [checked, setChecked] = React.useState(false);
-  const [status, setstatus] = React.useState<any>("");
-  const [agegrp, setagegrp] = React.useState<any>("");
-  const [custType, setcustType] = React.useState<any>("");
-  const [custAdd, setcustAdd] = React.useState<any>("");
+  const [custtype, setcusttype] = React.useState<any>([]);
+  const [AccId, setAccid] = React.useState(0);
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setselValue(newValue);
+    setValue(newValue);
   };
-  const [Check, setCheck] = React.useState(false);
-  if (Check === true) {
-    var hideshowstyle = {
-      display: "block",
-    };
-  } else {
-    var hideshowstyle = {
-      display: "none",
-    };
-  }
 
-  if (parentid > 0) {
-    var hideshowstyle = {
-      display: "block",
+  const {
+    register,
+    handleSubmit,
+    reset,
+    formState: { errors },
+  } = useForm<FormValues>();
+
+  const onSubmit: SubmitHandler<FormValues> = async (data) => {
+    const reqData = {
+      name: data.name,
+      email1: data.email1,
+      email2: data.email2,
+      phone1: data.phone1,
+      phone2: data.phone2,
+      contactName: data.contactName,
+      printUs: data.printUs,
+      status: data.status,
+      typeId: data.type,
+      generatedId: data.pregeneratedid,
+      parentId: 0,
+      userRole: "parent",
     };
-  }
+    await axios({
+      method: "POST",
+      url: `${api_url}/addUser`,
+      data: reqData,
+      headers: {
+        Authorization: auth_token,
+      },
+    })
+      .then((data) => {
+        if (data) {
+          setshowspinner(false);
+          setBtnDisabled(false);
+          toast.success("Parent Added successfully !");
+          reset();
+          closeDialog(false);
+        }
+      })
+      .catch((error) => {
+        toast.error("Email Allready Registred !");
+        setshowspinner(false);
+        setBtnDisabled(false);
+      });
+  };
 
   React.useEffect(() => {
     getType();
-    getUserDet();
-    getParentDet();
-    if (parentid > 0) {
-      setChecked(true);
-    }
-  }, [status]);
+    commmonfunctions.GetLastInsertId().then(res => {
+      setAccid(res.id);
+    })
+  }, []);
+
 
   //get type
   const getType = async () => {
@@ -185,143 +201,26 @@ export default function EditCustomer({
       console.log("error", error);
     }
   };
-  //get user details
-  const getUserDet = async () => {
-    try {
-      const response = await fetch(`${api_url}/getuserdetails/${id}`, {
-        method: "GET",
-        headers: {
-          Authorization: auth_token,
-        },
-      });
-      const res = await response.json();
-      setValue("name", res.data[0].name);
-      setValue("email1", res.data[0].email1);
-      setValue("email2", res.data[0].email2);
-      setValue("number", res.data[0].phone1);
-      setValue("alternatenumber", res.data[0].phone2 === 0 ? "" : res.data[0].phone2);
-      setValue("contactName", res.data[0].contactName);
-      setValue("printUs", res.data[0].printus);
-      setValue("pregeneratedid", res.data[0].generatedId);
-      setparentid(res.data[0].parentId);
-      setstatus(res.data[0].status);
-      setcustType(res.data[0].typeId);
-      setagegrp(res.data[0].agegroup);
-      setValue("pregeneratedid", res.data[0].generatedId);
-      const addr = JSON.parse(res.data[0]?.address);
-      setcustAdd(addr);
-      setValue("address1", custAdd?.add1);
-      setValue("address2", custAdd?.add2);
-      setValue("city", custAdd?.city);
-      setValue("state", custAdd?.state);
-      setValue("postalcode", custAdd?.postalcode);
-      setValue("attentionto", res.data[0]?.attentionTo);
-
-    } catch (error) {
-      console.log("error", error);
-    }
-  };
-  //get parent
-  const getParentDet = async () => {
-    try {
-      const response = await fetch(`${api_url}/getuserbypid/${parentid}`, {
-        method: "GET",
-        headers: {
-          Authorization: auth_token,
-        },
-      });
-      const res = await response.json();
-      setparentname(res.data[0].name);
-    } catch (error) {
-      console.log("error", error);
-    }
-  };
-
-  const {
-    register,
-    handleSubmit,
-    setValue,
-    formState: { errors },
-  } = useForm<FormValues>();
-
-  // submit data
-  const onSubmit: SubmitHandler<FormValues> = async (data) => {
-    setshowspinner(true);
-    setBtnDisabled(true);
-    const address = {
-      add1: data.address1,
-      add2: data.address2,
-      city: data.city,
-      state: data.state,
-      postalcode: data.postalcode,
-    };
-    const reqData = {
-      name: data.name,
-      email1: data.email1,
-      email2: data.email2,
-      phone1: data.number,
-      contactName: data.contactName,
-      printUs: data.printUs,
-      status: data.status == 1 ? "1" : data.status == 0 ? "0" : "",
-      typeId: data.customertype,
-      parentId: parentid,
-      agegroup: data.agegroup,
-      pregeneratedid: data.pregeneratedid,
-      updatedBy: 1,
-      attentionto: data.attentionto,
-      phone2: data.alternatenumber,
-      useraddress: address,
-    };
-    await axios({
-      method: "PUT",
-      url: `${api_url}/edituser/${id}`,
-      data: reqData,
-      headers: {
-        Authorization: auth_token,
-      },
-    })
-      .then((data) => {
-        if (data.status === 200) {
-          setshowspinner(false);
-          setBtnDisabled(false);
-          toast.success("Customer Updated Successfully !");
-          setTimeout(() => {
-            setOpen(false);
-          }, 2000);
-          closeDialogedit(data.status);
-        }
-      })
-      .catch((error) => {
-        toast.error("Email Allready Registred !");
-        setshowspinner(false);
-        setBtnDisabled(false);
-      });
-  };
 
   const closeDialogs = () => {
-    closeDialogedit(false);
-    setOpen(false);
-  };
-
-  const Getdata = (item: any) => {
-    setparentid(item && item.id);
+    closeDialog(false);
   };
 
   return (
     <BootstrapDialog
       onClose={closeDialogs}
       aria-labelledby="customized-dialog-title"
-      open={opens}
+      open={open}
     >
       <BootstrapDialogTitle id="customized-dialog-title" onClose={closeDialogs}>
-        Edit Customer
+        New Parent
       </BootstrapDialogTitle>
       <DialogContent dividers>
         <Box sx={{ width: "100%" }}>
           <form onSubmit={handleSubmit(onSubmit)}>
             <Box sx={{ borderBottom: 1, borderColor: "divider" }}>
               <Tabs
-                value={selvalue}
+                value={value}
                 onChange={handleChange}
                 aria-label="basic tabs example"
               >
@@ -330,31 +229,32 @@ export default function EditCustomer({
                 <Tab label="Options" {...a11yProps(2)} />
               </Tabs>
             </Box>
-            <TabPanel value={selvalue} index={0}>
+            <TabPanel value={value} index={0}>
               <Grid>
                 <Stack>
                   <Grid container spacing={2}>
                     <Grid item xs={12} md={12}>
                       <Stack spacing={1}>
                         <InputLabel htmlFor="name">
-                          Customer <span className="err_str">*</span>
+                          Parent Name <span className="err_str">*</span>
                         </InputLabel>
                         <OutlinedInput
                           type="text"
                           id="name"
-                          placeholder="Customer Name..."
+                          placeholder="Parent Name..."
                           fullWidth
                           size="small"
                           {...register("name", {
                             required: true,
                             validate: (value) => { return !!value.trim() }
                           })}
+                          defaultValue={name}
                         />
                         {errors.name?.type === "required" && (
                           <span style={style}>Field is Required *</span>
                         )}
                         {errors.name?.type === "validate" && (
-                          <span style={style}>Field can't be blank *</span>
+                          <span style={style}>Name can't be blank *</span>
                         )}
                       </Stack>
                       <FormGroup>
@@ -366,54 +266,22 @@ export default function EditCustomer({
                     </Grid>
                   </Grid>
                 </Stack>
-                <Stack style={{ marginTop: "8px" }}>
+                <Stack style={{ marginTop: "20px" }}>
                   <Grid container spacing={2}>
                     <Grid item xs={12} md={6}>
                       <Stack spacing={1}>
-                        <InputLabel htmlFor="name">
-                          Phone Number <span className="err_str">*</span>
-                        </InputLabel>
-                        <OutlinedInput
-                          type="text"
-                          id="name"
-                          placeholder="Phone..."
-                          fullWidth
-                          size="small"
-                          {...register("number", {
-                            required: true,
-                            pattern: /^[6-9]\d{9}$/,
-                            minLength: 10,
-                            maxLength: 10,
-                          })}
-                        />
-                        {errors.number?.type === "required" && (
-                          <span style={style}>Field is Required *</span>
-                        )}
-                        {errors.number?.type === "pattern" && (
-                          <span style={style}>Enter Valid Number *</span>
-                        )}
-                        {errors.number?.type === "minLength" && (
-                          <span style={style}>Enter Valid Number *</span>
-                        )}
-                        {errors.number?.type === "maxLength" && (
-                          <span style={style}>Enter Valid Number *</span>
-                        )}
-                      </Stack>
-                    </Grid>
-                    <Grid item xs={12} md={6}>
-                      <Stack spacing={1}>
-                        <InputLabel htmlFor="name">
+                        <InputLabel htmlFor="email">
                           Email <span className="err_str">*</span>
                         </InputLabel>
                         <OutlinedInput
                           type="text"
                           id="name"
-                          placeholder="Email..."
+                          placeholder="email..."
                           fullWidth
                           size="small"
                           {...register("email1", {
-                            required: true,
                             pattern: /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i,
+                            validate: (value) => { return !!value.trim() }
                           })}
                         />
                         {errors.email1?.type === "required" && (
@@ -424,36 +292,18 @@ export default function EditCustomer({
                             Please enter a valid email address *
                           </span>
                         )}
-                      </Stack>
-                    </Grid>
-                  </Grid>
-                </Stack>
-                <Stack style={{ marginTop: "15px" }}>
-                  <Grid container spacing={2}>
-                    <Grid item xs={12} md={6}>
-                      <Stack spacing={1}>
-                        <InputLabel htmlFor="name">Status</InputLabel>
-                        <FormControl fullWidth>
-                          {status !== "" ? (
-                            <Select
-                              labelId="demo-simple-select-label"
-                              id="demo-simple-select"
-                              size="small"
-                              {...register("status")}
-                              defaultValue={status}
-                            >
-                              <MenuItem value={1}>Active</MenuItem>
-                              <MenuItem value={0}>InActive</MenuItem>
-                            </Select>
-                          ) : (
-                            "loading......"
-                          )}
-                        </FormControl>
+                        {errors.email1?.type === "validate" && (
+                          <span style={style}>
+                            Please enter a valid email address *
+                          </span>
+                        )}
                       </Stack>
                     </Grid>
                     <Grid item xs={12} md={6}>
                       <Stack spacing={1}>
-                        <InputLabel htmlFor="name">Alternate Email</InputLabel>
+                        <InputLabel htmlFor="aemail">
+                          Alternate Email
+                        </InputLabel>
                         <OutlinedInput
                           type="text"
                           id="name"
@@ -466,9 +316,62 @@ export default function EditCustomer({
                     </Grid>
                   </Grid>
                 </Stack>
+                <Stack style={{ marginTop: "20px" }}>
+                  <Grid container spacing={2}>
+                    <Grid item xs={12} md={6}>
+                      <Stack spacing={1}>
+                        <InputLabel htmlFor="number">
+                          Phone Number <span className="err_str">*</span>
+                        </InputLabel>
+                        <OutlinedInput
+                          type="text"
+                          id="number"
+                          placeholder="number..."
+                          fullWidth
+                          size="small"
+                          {...register("phone1", {
+                            required: true,
+                            pattern: /^[0-9+-]+$/,
+                            minLength: 10,
+                            maxLength: 10,
+                          })}
+                        />
+                        {errors.phone1?.type === "required" && (
+                          <span style={style}>Field is Required *</span>
+                        )}
+                        {errors.phone1?.type === "pattern" && (
+                          <span style={style}>Enter Valid Number *</span>
+                        )}
+                        {errors.phone1?.type === "minLength" && (
+                          <span style={style}>Enter Valid Number *</span>
+                        )}
+                        {errors.phone1?.type === "maxLength" && (
+                          <span style={style}>Enter Valid Number *</span>
+                        )}
+                      </Stack>
+                    </Grid>
+                    <Grid item xs={12} md={6}>
+                      <Stack spacing={1}>
+                        <InputLabel htmlFor="name">Status</InputLabel>
+                        <FormControl fullWidth>
+                          <Select
+                            labelId="demo-simple-select-label"
+                            id="demo-simple-select"
+                            defaultValue={1}
+                            size="small"
+                            {...register("status")}
+                          >
+                            <MenuItem value={1}>Active</MenuItem>
+                            <MenuItem value={0}>InActive</MenuItem>
+                          </Select>
+                        </FormControl>
+                      </Stack>
+                    </Grid>
+                  </Grid>
+                </Stack>
               </Grid>
             </TabPanel>
-            <TabPanel value={selvalue} index={1}>
+            <TabPanel value={value} index={1}>
               <Grid>
                 <Stack style={{ marginTop: "20px" }}>
                   <Grid container spacing={2}>
@@ -478,10 +381,9 @@ export default function EditCustomer({
                         <OutlinedInput
                           type="text"
                           id="name"
-                          placeholder="Attention To..."
+                          placeholder="Attention To...."
                           fullWidth
                           size="small"
-                          {...register("attentionto")}
                         />
                       </Stack>
                     </Grid>
@@ -491,10 +393,9 @@ export default function EditCustomer({
                         <OutlinedInput
                           type="text"
                           id="name"
-                          placeholder="Phone..."
+                          placeholder="Phone...."
                           fullWidth
                           size="small"
-                          {...register("alternatenumber")}
                         />
                       </Stack>
                     </Grid>
@@ -508,10 +409,9 @@ export default function EditCustomer({
                         <OutlinedInput
                           type="text"
                           id="name"
-                          placeholder="Address1..."
+                          placeholder="Address 1..."
                           fullWidth
                           size="small"
-                          {...register("address1")}
                         />
                       </Stack>
                     </Grid>
@@ -521,10 +421,9 @@ export default function EditCustomer({
                         <OutlinedInput
                           type="text"
                           id="name"
-                          placeholder="Address2..."
+                          placeholder="Address 2..."
                           fullWidth
                           size="small"
-                          {...register("address2")}
                         />
                       </Stack>
                     </Grid>
@@ -541,7 +440,6 @@ export default function EditCustomer({
                           placeholder="City..."
                           fullWidth
                           size="small"
-                          {...register("city")}
                         />
                       </Stack>
                     </Grid>
@@ -554,7 +452,6 @@ export default function EditCustomer({
                           placeholder="State..."
                           fullWidth
                           size="small"
-                          {...register("state")}
                         />
                       </Stack>
                     </Grid>
@@ -567,7 +464,6 @@ export default function EditCustomer({
                           placeholder="Postal Code..."
                           fullWidth
                           size="small"
-                          {...register("postalcode")}
                         />
                       </Stack>
                     </Grid>
@@ -575,20 +471,20 @@ export default function EditCustomer({
                 </Stack>
               </Grid>
             </TabPanel>
-            <TabPanel value={selvalue} index={2}>
+            <TabPanel value={value} index={2}>
               <Grid>
                 <Stack>
                   <Grid container spacing={2}>
                     <Grid item xs={12} md={12}>
                       <Stack spacing={1}>
-                        <InputLabel htmlFor="name">Customer Type</InputLabel>
+                        <InputLabel htmlFor="name">Parent Type</InputLabel>
                         <FormControl fullWidth>
                           <Select
                             labelId="demo-simple-select-label"
                             id="demo-simple-select"
                             size="small"
-                            defaultValue={custType}
-                            {...register("customertype")}
+                            defaultValue={0}
+                            {...register("type")}
                           >
                             <MenuItem value={0}>Individual</MenuItem>
                             {custtype &&
@@ -605,27 +501,21 @@ export default function EditCustomer({
                     </Grid>
                   </Grid>
                 </Stack>
-                <Stack style={{ marginTop: "10px" }}>
-                  <Grid container spacing={2}>
+                <Stack>
+                  <Grid container spacing={2} style={{ marginTop: "15px" }}>
                     <Grid item xs={12} md={12}>
-                      <FormGroup>
-                        <FormControlLabel
-                          control={
-                            <Checkbox
-                              onChange={(e) => setCheck(e.target.checked)}
-                              defaultChecked={checked}
-                            />
-                          }
-                          label="Belongs to a parent customer"
-                        />
-                      </FormGroup>
-                      <Stack spacing={1} style={hideshowstyle}>
-                        <InputLabel htmlFor="name"></InputLabel>
-                        <AddCustomerCmp
-                          Data={Getdata}
-                          PId={parentid}
-                          pname={parentname}
-                        />
+                      <Stack spacing={1}>
+                        <InputLabel htmlFor="name">Alternate Number</InputLabel>
+                        <FormControl fullWidth>
+                          <OutlinedInput
+                            type="text"
+                            id="name"
+                            placeholder="Contact Name..."
+                            fullWidth
+                            size="small"
+                            {...register("phone2")}
+                          />
+                        </FormControl>
                       </Stack>
                     </Grid>
                   </Grid>
@@ -643,10 +533,13 @@ export default function EditCustomer({
                           placeholder="Contact Name..."
                           fullWidth
                           size="small"
-                          {...register("contactName", { required: true })}
+                          {...register("contactName", { required: true, validate: (value) => { return !!value.trim() } })}
                         />
-                        {errors.contactName && (
+                        {errors.contactName?.type === "required" && (
                           <span style={style}>Field is Required *</span>
+                        )}
+                        {errors.contactName?.type === "validate" && (
+                          <span style={style}>Field can't be blank *</span>
                         )}
                       </Stack>
                     </Grid>
@@ -661,49 +554,34 @@ export default function EditCustomer({
                           placeholder="Print Us..."
                           fullWidth
                           size="small"
-                          {...register("printUs", { required: true })}
+                          {...register("printUs", { required: true, validate: (value) => { return !!value.trim() } })}
                         />
-                        {errors.printUs && (
+                        {errors.printUs?.type === "required" && (
                           <span style={style}>Field is Required *</span>
+                        )}
+                        {errors.printUs?.type === "validate" && (
+                          <span style={style}>Field can't be blank *</span>
                         )}
                       </Stack>
                     </Grid>
                   </Grid>
                 </Stack>
-                <Stack style={{ marginTop: "15px" }}>
-                  <Grid container spacing={2}>
-                    <Grid item xs={12} md={12}>
-                      <Stack spacing={1}>
-                        <InputLabel htmlFor="name">Age Group</InputLabel>
-                        <FormControl fullWidth>
-                          {status !== "" ? (
-                            <Select
-                              labelId="demo-simple-select-label"
-                              id="demo-simple-select"
-                              defaultValue={agegrp}
-                              size="small"
-                              {...register("agegroup")}
-                            >
-                              <MenuItem value={1}>FS1</MenuItem>
-                              <MenuItem value={2}>FS2</MenuItem>
-                              <MenuItem value={3}>FS3</MenuItem>
-                              <MenuItem value={4}>FS4</MenuItem>
-                              <MenuItem value={5}>FS5</MenuItem>
-                              <MenuItem value={6}>FS6</MenuItem>
-                              <MenuItem value={7}>FS7</MenuItem>
-                              <MenuItem value={8}>FS8</MenuItem>
-                              <MenuItem value={9}>FS9</MenuItem>
-                              <MenuItem value={10}>FS10</MenuItem>
-                              <MenuItem value={11}>FS11</MenuItem>
-                              <MenuItem value={12}>FS12</MenuItem>
-                              <MenuItem value={13}>FS13</MenuItem>
-                            </Select>) : ("loadinf...")
-                          }
-                        </FormControl>
-                      </Stack>
-                    </Grid>
-                  </Grid>
-                </Stack>
+                <Grid item xs={12} md={12} style={{ marginTop: "10px" }}>
+                  <Stack spacing={1}>
+                    <InputLabel htmlFor="name">
+                      Account
+                    </InputLabel>
+                    <OutlinedInput
+                      type="text"
+                      id="name"
+                      placeholder="# Generate If blank"
+                      fullWidth
+                      value={`#CUST${AccId && AccId + 1}`}
+                      size="small"
+                      {...register("pregeneratedid")}
+                    />
+                  </Stack>
+                </Grid>
               </Grid>
             </TabPanel>
             <DialogActions>
@@ -715,15 +593,15 @@ export default function EditCustomer({
                 autoFocus
                 disabled={btnDisabled}
               >
-                <b>Update</b>
+                <b>Create</b>
                 <span style={{ fontSize: "2px", paddingLeft: "10px" }}>
                   {spinner === true ? <CircularProgress color="inherit" /> : ""}
                 </span>
               </Button>
             </DialogActions>
           </form>
-          <ToastContainer />
         </Box>
+        <ToastContainer />
       </DialogContent>
     </BootstrapDialog>
   );
