+++ pages/admin/customer/editType.tsx
@@ -15,7 +15,7 @@ import React, { useEffect } from "react";
 import axios from "axios";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import { api_url, auth_token } from "../api/hello";
+import { api_url, auth_token } from "../../api/hello";
 import styled from "@emotion/styled";
 import { GridCloseIcon } from "@mui/x-data-grid";
 import { useForm, SubmitHandler } from "react-hook-form";
