+++ pages/commoncmp/addCustomerCmp.tsx
@@ -2,7 +2,7 @@ import { Autocomplete, Button, TextField } from "@mui/material";
 import React, { useEffect, useState } from "react";
 
 import { api_url, auth_token } from "../api/hello";
-import AddNewParent from "../customer/addNewParent";
+import AddNewParent from "../admin/customer/addNewParent";
 
 export default function AddCustomerCmp({ Data, PId, pname }: { Data: any, PId: any, pname: any }) {
   const [users, setUsers] = useState<any>([]);
