+++ pages/user/creditinvoices/creditinvoiceview/[id].tsx
@@ -19,6 +19,7 @@ import { api_url, auth_token } from "../../../api/hello";
 import { useRouter } from "next/router";
 import MainFooter from "../../../commoncmp/mainfooter";
 import moment from "moment";
+import commmonfunctions from "../../../../commonFunctions/commmonfunctions";
 
 export default function ViewCreditNotes(props: any) {
     const [creditNoteDet, setcreditNoteDet] = React.useState<any>([]);
@@ -31,13 +32,23 @@ export default function ViewCreditNotes(props: any) {
     // verify user login and previlegs
     let logintoken: any;
     useEffect(() => {
+        commmonfunctions.VerifyLoginUser().then(res => {
+            if (res.exp * 1000 < Date.now()) {
+                localStorage.removeItem('QIS_loginToken');
+            }
+        });
         logintoken = localStorage.getItem("QIS_loginToken");
         if (logintoken === undefined || logintoken === null) {
             router.push("/");
         }
+        commmonfunctions.GivenPermition().then((res) => {
+            if (res.roleId === 2) {
+            } else {
+                router.push("/");
+            }
+        });
         fetchData();
     }, []);
-
 
     //get credit notes
     const url = `${api_url}/getCreditNotesDetails/${id}`;
@@ -74,9 +85,6 @@ export default function ViewCreditNotes(props: any) {
             console.log("error", error);
         }
     };
-
-    console.log(creditNoteDet);
-
     return (
         <>
             <Box sx={{ display: "flex" }}>
