+++ pages/user/dashboard.tsx
@@ -38,6 +38,7 @@ import DialogActions from "@mui/material/DialogActions";
 import IconButton from "@mui/material/IconButton";
 import CloseIcon from "@mui/icons-material/Close";
 import { useForm, SubmitHandler } from "react-hook-form";
+import commmonfunctions from "../../commonFunctions/commmonfunctions";
 
 const style = {
   color: "red",
@@ -80,6 +81,7 @@ type FormValues = {
   state: string;
   postalcode: number;
 };
+
 export interface DialogTitleProps {
   id: string;
   children?: React.ReactNode;
@@ -88,7 +90,6 @@ export interface DialogTitleProps {
 
 function BootstrapDialogTitle(props: DialogTitleProps) {
   const { children, onClose, ...other } = props;
-
   return (
     <DialogTitle sx={{ m: 0, p: 2 }} {...other}>
       {children}
@@ -171,7 +172,22 @@ export default function ViewCustomer() {
 
   const verifyLoginUser = async () => {
     let login_token: any;
+    commmonfunctions.VerifyLoginUser().then(res => {
+      if (res.exp * 1000 < Date.now()) {
+        localStorage.removeItem('QIS_loginToken');
+        router.push("/");
+      }
+    });
     login_token = localStorage.getItem("QIS_loginToken");
+    if (login_token === undefined || login_token === null) {
+      router.push("/");
+    }
+    commmonfunctions.GivenPermition().then(res => {
+      if (res.roleId !== 2) {
+        router.push("/");
+      }
+    });
+
     const decoded: any = jwt_decode(login_token);
     setCustomerId(decoded?.id);
     fetchBallance(decoded?.id);
